third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/sub-pixel/inline-block-with-padding-expected.txt

PASS links[0].offsetWidth is within 5 of 175
PASS successfullyParsed is true

TEST COMPLETE


Both links should render the same and not wrap.

This shouldn't wrap
This shouldn't wrap
